requirements.txt
frida
prompt_toolkit
click
jinja2
tabulate
delegator.py
requests